edited order classes

+++ src/main/java/com/example/infraboxapi/order/OrderService.java
@@ -5,13 +5,11 @@ import com.example.infraboxapi.notification.NotificationDescription;
 import com.example.infraboxapi.notification.NotificationService;
 import com.example.infraboxapi.orderItem.OrderItem;
 import com.example.infraboxapi.orderItem.OrderItemDTO;
-import com.example.infraboxapi.orderItem.OrderItemRepository;
 import com.example.infraboxapi.tool.Tool;
 import com.example.infraboxapi.tool.ToolRepository;
 import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
-
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -21,7 +19,6 @@ import java.util.Optional;
 public class OrderService {
 
     private final OrderRepository orderRepository;
-    private final OrderItemRepository orderItemRepository;
     private final MaterialRepository materialRepository;
     private final ToolRepository toolRepository;
     private final NotificationService notificationService;
@@ -32,37 +29,11 @@ public class OrderService {
 
     @Transactional
     public void addOrder(OrderDTO orderDTO) {
-
-
-
         List<OrderItem> orderItems = new ArrayList<>();
 
         for (OrderItemDTO orderItemDTO : orderDTO.getOrderItems()) {
-
-
-            if(orderItemDTO.getItemType().equals("tool")){
-                OrderItem orderItem = OrderItem.builder()
-                        .name(orderItemDTO.getName())
-                        .quantity(orderItemDTO.getQuantity())
-                        .material(null)
-                        .tool(toolRepository.findById(orderItemDTO.getItemID()).orElse(null))
-                        .build();
-
-                orderItems.add(orderItem);
-            }else{
-                OrderItem orderItem = OrderItem.builder()
-                        .name(orderItemDTO.getName())
-                        .quantity(orderItemDTO.getQuantity())
-                        .material(materialRepository.findById(orderItemDTO.getItemID()).orElse(null))
-                        .tool(null)
-                        .build();
-
-                orderItems.add(orderItem);
-            }
-
-
-
-
+            OrderItem orderItem = createOrderItemFromDTO(orderItemDTO);
+            orderItems.add(orderItem);
         }
 
         Order order = Order.builder()
@@ -75,21 +46,44 @@ public class OrderService {
                 .orderItems(orderItems)
                 .build();
 
-
-
         orderRepository.save(order);
 
-        notificationService.createAndSendNotification("Order  " + order.getName() + " has been added.", NotificationDescription.OrderAdded);
-
-
+        notificationService.createAndSendNotification("Order " + order.getName() + " has been added.", NotificationDescription.OrderAdded);
     }
 
+    private OrderItem createOrderItemFromDTO(OrderItemDTO orderItemDTO) {
+        String itemType = orderItemDTO.getItemType();
+        OrderItem.OrderItemBuilder orderItemBuilder = OrderItem.builder()
+                .name(orderItemDTO.getName())
+                .quantity(orderItemDTO.getQuantity());
+
+        if ("tool".equals(itemType)) {
+            orderItemBuilder.tool(toolRepository.findById(orderItemDTO.getItemID()).orElse(null));
+        } else {
+            orderItemBuilder.material(materialRepository.findById(orderItemDTO.getItemID()).orElse(null));
+        }
+
+        return orderItemBuilder.build();
+    }
+
+    @Transactional
     public void deleteOrder(Integer id) {
-        setQuantityInTransportToZero(id);
-        orderRepository.deleteById(id);
+        Optional<Order> orderOptional = orderRepository.findById(id);
+
+        if (orderOptional.isPresent()) {
+            Order order = orderOptional.get();
+            if ("on the way".equals(order.getStatus())) {
+                updateQuantityInTransport(id);
+            }
+            orderRepository.deleteById(id);
+            notificationService.createAndSendNotification("Order  " + order.getName() + " has been deleted.", NotificationDescription.OrderDeleted);
+        }
+
+
+
     }
 
-    public void setQuantityInTransportToZero(Integer id) {
+    public void updateQuantityInTransport(Integer id) {
         Optional<Order> orderOptional = orderRepository.findById(id);
 
         if (orderOptional.isPresent()) {
@@ -100,7 +94,12 @@ public class OrderService {
                     Material material = materialRepository.findById(orderItem.getMaterial().getId()).orElse(null);
 
                     if (material != null) {
-                        material.setQuantityInTransit(0);
+                        if(material.getQuantityInTransit() - orderItem.getQuantity() < 0){
+                            material.setQuantityInTransit(0);
+                        }else{
+                            material.setQuantityInTransit(material.getQuantityInTransit() - orderItem.getQuantity());
+                        }
+
                         materialRepository.save(material);
                     }
                 } else if (orderItem.getTool() != null) {
@@ -108,7 +107,12 @@ public class OrderService {
                     Tool tool = toolRepository.findById(orderItem.getTool().getId()).orElse(null);
 
                     if (tool != null) {
-                        tool.setQuantityInTransit(0);
+                        if(tool.getQuantityInTransit() - orderItem.getQuantity() < 0){
+                            tool.setQuantityInTransit(0);
+                        }else{
+                            tool.setQuantityInTransit(tool.getQuantityInTransit() - orderItem.getQuantity());
+                        }
+
                         toolRepository.save(tool);
                     }
                 }
